Suspend signals during authentication with token files.
This fixes a security bug where a user who is permitted using the
'persist' configuration option can bypass password entry. If doas
catches a signal such as SIGINT (Ctrl-C) at a password prompt and
also has an auth token file open, then doas is terminated without
first clearing the token file. This means that doas can then be rerun
and the user will be "authenticated" due to the token file present.
Blocking certain signals while an auth token file is open solves
this problem (or at least *should*).

The signals that are blocked are those handled by
bsd-compat/readpassphrase.c, which seems like a good set to start
with. Console interrupts are the most immediately concerning, and
we shouldn't need to worry about other signals which could cause
termination because doas is a root-owned process when running.
(Yes, this means doas can be interrupted by another root-owned
process, but to do that you have to *get* root in the first place,
against which this commit defends.)