Refactor timestamp file handling code.
This is a fairly substantial change to the timestamp handling code, which
includes: NULL termination handling (from Duncan Overbruck's opendoas);
dropping redundant ioctl()s to discover the session leader in favour of just
calling getsid() (as the ioctl() and getsid() should always return the same
value); handle the case where a new timestamp file is created (instead of
updating an existing one) specially, by first creating a temporary file, and
then renaming it if authentication succeeds.

The last change requires tracking a little more state in the caller, as the
generate timestamp file name and temporary file name are needed in order to
call renameat() in persist_commit().